+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## unreleased
 
+## 2.1.0 - 2026-06-12
+
 ### Added
 
  - The `checked_oint.guard` virtual library for controlling the behaviour of polymorphic comparison, with two implementations: